Remove compatility with Windows
The text art does not print on Windows due to problems with encoding.
Until I come up with a sustainable solution to fix that -- or decide to make
the text art so that it only contains characters Windows can print too --
I'll declare that Commandsheet is compatible with MacOS and Linux only.